+++ platformio-project/lib/DTSDataCollector/DTSDataCollector.cpp
@@ -33,7 +33,7 @@ DTSDataCollector::DTSDataCollector(std::string channelName, uint8_t oneWirePin)
 /// Constructing with a channelName 
 /// a positive or degative temperatureOffset to adjust sensor tolerances
 /// a onewire-pin
-DTSDataCollector::DTSDataCollector(std::string channelName, double temperatureOffset,uint8_t oneWirePin)
+DTSDataCollector::DTSDataCollector(std::string channelName, float temperatureOffset,uint8_t oneWirePin)
 {
     _channelName = channelName;
     _oneWirePin = oneWirePin;
@@ -64,7 +64,7 @@ void DTSDataCollector::updateData()
     #ifdef DTSDATACOLLECTOR_H_DEBUG
         Serial.println("DEBUG CODE ACTIVE! RANDOM DATA");
         long randomVal = random(-200,400);
-        double result = randomVal/10;
+        float result = randomVal/10;
         _dataStorage->addData(_channelName + CHANNELEXTTEMP, result);
     #else
         // #rb: Der getDeviceCount führt dazu, dass die FW den Sensor nicht findet
